+++ packages/logic-core/src/data/solver/backtrack/backtrackSolver.ts
@@ -15,6 +15,7 @@ import { instance as focusInstance } from '../../symbols/focusSymbol.js';
 import { instance as myopiaInstance } from '../../symbols/myopiaSymbol.js';
 import { instance as viewpointInstance } from '../../symbols/viewpointSymbol.js';
 import { instance as connectAllInstance } from '../../rules/connectAllRule.js';
+import { instance as unsupportedInstance } from '../../symbols/unsupportedSymbol.js';
 import EventIteratingSolver from '../eventIteratingSolver.js';
 import Instruction from '../../instruction.js';
 import GridData from '../../grid.js';
@@ -41,6 +42,7 @@ export default class BacktrackSolver extends EventIteratingSolver {
     cellCountInstance.id,
     sameShapeInstance.id,
     uniqueShapeInstance.id,
+    unsupportedInstance.id,
   ];
 
   public readonly id = 'backtrack';